src/test/demoProject/requirements.txt
numpy
torch
tensorflow
pillow
jax